requirements.txt
streamlit
folium
streamlit-folium
plotly
pandas
numpy
ultralytics
line-bot-sdk
flask
python-dotenv
google-genai
